Change implementation of gradient on iOS
`CAGradientLayer` has numerous issues with accuracy of the produced
gradients for non-square views. This changes the implementation to a
custom layer with `CGContextDrawLinearGradient`. `drawInContext:` is
implemented, but is called manuall from `display` which ensures that
React Native's custom code for rendering corner radii etc still works
as expected. However with some complex React Native view properties the
implementation will not work, but this is not a regression from the
`CAGradientLayer` implementation.

This is based on previous work done by
[danielgindi](https://github.com/danielgindi) in
https://github.com/react-native-community/react-native-linear-gradient/pull/215.